Split module Console into several 'role' modules
Create PromptReader, PromptWriter and OptionsDisplay
